// src/models/FirewallGroupOutResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FirewallGroupOutAPI } from './FirewallGroupOutAPI';
export type FirewallGroupOutResponse = {
    /**
     * Идентификатор запроса
     */
    response_id?: string;
    /**
     * Группа правил
     */
    group: FirewallGroupOutAPI;
};