// src/models/FirewallGroupResourceOutResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FirewallGroupResourceOutAPI } from './FirewallGroupResourceOutAPI';
export type FirewallGroupResourceOutResponse = {
    /**
     * Идентификатор запроса
     */
    response_id?: string;
    resource: FirewallGroupResourceOutAPI;
};