// src/models/FirewallRuleOutResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FirewallRuleOutAPI } from './FirewallRuleOutAPI';
export type FirewallRuleOutResponse = {
    /**
     * Идентификатор запроса
     */
    response_id?: string;
    /**
     * Firewall правило
     */
    rule: FirewallRuleOutAPI;
};